Move tests into test file, maybe that'll do it?
